pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agent1"
version = "0.2.0"
description = "Hierarchical agent system based on DeepResearchAgent patterns"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "claude-agent-sdk>=0.1.5",
    "hydra-core>=1.3.2",
    "omegaconf>=2.3.0",
    "pandas>=2.0.0",
    "nest-asyncio>=1.5.0",
    "rich>=13.0.0",
    "pyyaml>=6.0.0",
    "jinja2>=3.1.0",
    "aiofiles>=23.0.0",
    "asyncio>=3.4.3",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
]

[tool.hatch.build.targets.wheel]
packages = ["agent1", "examples", "configs"]
